refacor: split paths and settings into dedicated modules
