+++ Cargo.toml
@@ -17,8 +17,7 @@ A dotfile manager.
 config = "0.11.0"
 dirs = "^4"
 toml = "^0"
-serde_derive = "^1"
-serde = "^1"
+serde = { version = "1.0.136", features = ["derive"] }
 anyhow = "^1"
 tera = "1"
 colored = "^2"
@@ -29,6 +28,7 @@ clap_complete = { version = "3.0", optional = true }
 lazy_static = { version = "1.4.0", optional = true }
 watchexec = { version = "2.0.0-pre.14"}
 tokio = { version = "1", features = ["macros", "rt"] }
+thiserror = "1.0.30"
 
 [features]
 default = ["cli"]
