+++ src/hook.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, Result};
 use colored::*;
-use serde_derive::{Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
 use std::io::{BufRead, BufReader};
 use std::process::{Command, Stdio};
